+++ src/main/java/com/techelevator/products/Chips.java
@@ -4,7 +4,7 @@ package com.techelevator.products;
 import java.math.BigDecimal;
 
 public class Chips extends Inventory {
-    public Chips(String vendPosition, String productName, double price, String type) {
+    public Chips(String vendPosition, String productName, BigDecimal price, String type) {
         super(vendPosition, productName, price, type);
     }
 }
