+++ src/main/java/com/techelevator/products/Drink.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 public class Drink extends Inventory {
 
-    public Drink(String vendPosition, String productName, double price, String type) {
+    public Drink(String vendPosition, String productName, BigDecimal price, String type) {
         super(vendPosition, productName, price, type);
     }
 }
